Restructured themes to use default, added styles files, added favicon, updated tests
