--- src/App.test.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-import { render } from '@testing-library/react';
-import App from './App';
-
-test('renders app', () => {
-  render(<App />);
-});
