+++ src/components/FormContent/styles.js
@@ -1,0 +1,12 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export default makeStyles(
+  theme => {
+    return {
+      root: {
+        padding: theme.variables.spacing.globalPadding
+      }
+    };
+  },
+  { name: 'FormContent' }
+);
