+++ src/components/FormSectionHeader/__tests__/FormSectionHeader.test.jsx
@@ -1,10 +1,17 @@
 import React from 'react';
 import { render } from '@testing-library/react';
+import { ThemeProvider } from '@material-ui/core/styles';
+
+import theme from '../../../styles/theme';
 import FormSectionHeader from '../FormSectionHeader';
 
 test('renders form section header with title', () => {
   const mockTitle = 'Example';
-  const { getByText } = render(<FormSectionHeader title={mockTitle} />);
+  const { getByText } = render(
+    <ThemeProvider theme={theme}>
+      <FormSectionHeader title={mockTitle} />
+    </ThemeProvider>
+  );
 
   const titleElement = getByText(mockTitle);
   expect(titleElement).toBeInTheDocument();
