+++ src/components/Navigation/Navigation.jsx
@@ -1,53 +1,13 @@
 import React, { useState, useCallback } from 'react';
 import clsx from 'clsx';
-import { makeStyles } from '@material-ui/core/styles';
 import tocbot from 'tocbot';
 
-const useStyles = makeStyles(theme => ({
-  toc: {
-    background: theme.palette.common.gray,
-    height: '100vh',
-    position: 'sticky',
-    top: '0',
-    paddingRight: '1em',
-    width: '265px',
-    float: 'left',
-    '@global': {
-      ol: {
-        listStyle: 'none',
-        fontFamily: theme.typography.fontFamily,
-        fontWeight: 600,
-        padding: '1.5em',
-        margin: '0',
-        textTransform: 'capitalize',
-        '@global': {
-          li: {
-            padding: '0.25em 0',
-            borderLeft: '2px solid #eee',
-            paddingLeft: '1em',
-
-            '&.is-active-li': {
-              borderColor: theme.palette.common.red,
-              fontWeight: 700
-            }
-          },
-          a: {
-            color: theme.palette.text.primary,
-            textDecoration: 'none'
-          }
-        }
-      }
-    }
-  },
-  makeSticky: {
-    position: 'fixed',
-    top: '0'
-  }
-}));
+import useStyles from './styles';
 
 function Navigation() {
   const styles = useStyles();
   const [classNames, setClassNames] = useState(clsx('toc', styles.toc));
+
   const scrollListener = useCallback(
     element => {
       const distanceToTop = element?.getBoundingClientRect().top;
@@ -63,6 +23,7 @@ function Navigation() {
     },
     [styles]
   );
+
   return <div ref={scrollListener} className={classNames}></div>;
 }
