+++ src/components/PageHeader/PageHeader.jsx
@@ -1,25 +1,6 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
 
-import Image from './covid.png';
-
-const useStyles = makeStyles(theme => ({
-  bg: {
-    backgroundImage: `url(${Image})`,
-    backgroundSize: '100%',
-    backgroundRepeat: 'no-repeat',
-    alignItems: 'center'
-  },
-  text: {
-    color: 'white',
-    fontFamily: 'Roboto, sans-serif',
-    fontWeight: 300,
-    fontSize: 'calc(12px + 0.9vw)',
-    padding: 'calc(2px + 1vw) 40px',
-    width: '50%',
-    minWidth: '500px'
-  }
-}));
+import useStyles from './styles';
 
 function PageHeader() {
   const styles = useStyles();
