--- src/components/ThemeProvider/index.js
@@ -1,2 +1,0 @@
-export { default } from './ThemeProvider';
-export { useTheme } from './ThemeProvider';
